chore(serve): add new parseArgs helper
